Implement heartbeat DDS multiplier and fix ghost sync sentinel logic [Performance] Replace spontaneous firing period division with precomputed DDS multiplier * Replace `spontaneous_firing_period_ticks` with precomputed `heartbeat_m` (65536 / period) in VariantParameters struct across CUDA, HIP, and Rust layouts * Update neuron update kernels in physics.cu, physics.hip, and cpu/physics.rs to use branchless DDS phase calculation: `phase = (current_tick * heartbeat_m + tid * 104729) & 0xFFFF` * Add heartbeat_m field serialization in axicor-baker biology.rs, main.rs, and parser/blueprints.rs with zero-cost precomputation * Propagate heartbeat_m through ManifestVariant in axicor-core config/manifest.rs with fallback calculation
[System] Fix ghost sync kernel sentinel condition
* Change sentinel skip condition from `head >= 0x70000000u` to exact match `head == 0x80000000u` in cu_ghost_sync_kernel (CUDA and HIP)
* Ensure fresh spikes (0xFFFFFFFF) pass through the barrier correctly

[Documentation] Update Ant-v4 example README with corrected commands and details
* Rename title from "Genesis HFT: Ant-v4 Example" to "Axicor HFT: Ant-v4 Connectome"
* Update all CLI commands from `genesis-node` to `axicor-node` and correct artifact path to `Axicor-Models/AntConnectome.axic`
* Elaborate on architectural principles, build steps, and troubleshooting for UDP client connection

Refs: DOD FIX annotations for zero-cost precomputation and sentinel correction

+++ axicor-node/src/node/shard_thread.rs
@@ -453,7 +453,7 @@ fn execute_night_phase(
                         axicor_compute::ffi::gpu_memcpy_host_to_device(
                             gpu.vram.ptrs.dendrite_weights as *mut _,
                             workspace.weights_slice_mut(padded_n).as_ptr() as *const _,
-                            dendrites_count * std::mem::size_of::<i16>(),
+                            dendrites_count * std::mem::size_of::<i32>(), // Закон Mass Domain (i32)
                         );
                         axicor_compute::ffi::gpu_memcpy_host_to_device(
                             gpu.vram.ptrs.soma_voltage as *mut _,
